core/pyproject.toml
[project]
name = "core"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "python-jose[cryptography]>=3.3",
]

[project.optional-dependencies]
sqlite = ["pysqlite3-binary; platform_system=='Windows'"]
postgres = ["psycopg[binary]>=3.2"]